requirements.txt
Django==5.0.2
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
django-cors-headers==4.3.1
psycopg2-binary==2.9.9
python-dotenv==1.0.1
firebase-admin==6.4.0
django-redis==5.4.0
Pillow==10.2.0
django-storages==1.14.2
pytest==8.0.2
pytest-django==4.8.0
coverage==7.4.3 